github_repohunter/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
httpx
chromadb
sentence-transformers
torch
unsloth
